add try ~ catch phrase at CalculatorMainLv3

+++ src/level3/operation/AbstractOperation.java
@@ -1,5 +1,5 @@
 package level3.operation;
 
 public abstract class AbstractOperation {
-    public abstract double operate(int num1, int num2);
+    public abstract double operate(double num1, double num2);
 }
