+++ src/level3/operation/AddOperation.java
@@ -2,7 +2,7 @@ package level3.operation;
 
 public class AddOperation extends AbstractOperation {
     @Override
-    public double operate(int num1, int num2) {
-        return (double) num1 + num2;
+    public double operate(double num1, double num2) {
+        return num1 + num2;
     }
 }
